Adding some stuffs: polar perspective. Fixing bugs to non-points. Refatoring and ajust norm

+++ srcs/ft_vectors.c
@@ -6,7 +6,7 @@
 /*   By: tlavared <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/09/16 11:26:36 by tlavared          #+#    #+#             */
-/*   Updated: 2025/10/01 01:43:06 by tlavared         ###   ########.fr       */
+/*   Updated: 2025/10/11 00:55:39 by tlavared         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -33,6 +33,6 @@ t_vec2	ft_get2d(t_fdf *f, int x, int y)
 	point2d = ft_iso(point3d);
 	point2d.x = (int)(point2d.x * f->scale + (float )f->offset_x);
 	point2d.y = (int)(point2d.y * f->scale + (float )f->offset_y);
-	point2d.color = f->map.colors[y][x];
+	point2d.color = (uint32_t) f->map.colors[y][x];
 	return (point2d);
 }
